pyproject.toml
[project]
name = "decp-processing"
description = "An introductory project to Prefect"
version = "0.1.0"
requires-python = ">= 3.8"
authors = [
    { name = "Colin Maudry", email = "colin+decp@maudry.com" }
]
dependencies = [
    "python-dotenv",
    "pandas",
    "pyarrow",
    "tableschema",
    "frictionless",
    "datapackage-to-datasette",
    "ipykernel",
    "prefect",
    "json_stream"
]

[project.optional-dependencies]
dev = [
    "pre-commit",
    "pytest-env",
    "pytest",
]

[tool.pytest.ini_options]
pythonpath = [
  "src"
]
testpaths = [
    "tests",
]
env = [
    "DECP_ENRICHIES_VALIDES_2019_URL=data/decp_augmente_valides_2019_test.csv",
    "DECP_ENRICHIES_VALIDES_2022_URL=data/decp_augmente_valides_2022_test.csv"
]
addopts = "-p no:warnings"
